Prompt user for unsaved changes.  Closes #2
The user is thrown back into the vim session if there are unsaved
changes.  At this point ranger should have exited, and the detach key
mapping should have been removed, so the user will have to exit vim
manually.  This may cause some confusion if the user decides to continue
editing, but this is probably better than clobbering unsaved changes.